refactor: Coupon 타입을 Coupon 컴포넌트 디렉터리에서 관리하도록 분리
